Add memory_usage module to sparse_set
